move internal reference
