+++ src/type_sys/fender_value/iterator.rs
@@ -4,7 +4,7 @@ use std::fmt::Debug;
 use std::rc::Rc;
 
 use crate::type_sys::{
-    fender_reference::FenderReference, fender_value::FenderValue, type_system::FenderTypeSystem,
+    fender_reference::FenderReference, fender_value::FenderValue, freight_type_system::FenderTypeSystem,
 };
 
 pub type NativeClosure =
