+++ src/type_sys/freight_type_system.rs
@@ -3,13 +3,14 @@ use super::{
     type_id::FenderTypeId,
 };
 use crate::{
+    fender_value::FenderValue,
     operators::FenderBinaryOperator,
     operators::FenderInitializer,
     operators::FenderUnaryOperator,
     plugin::PluginManager,
     stdlib::{loader::DependencyList, FenderResource, STDLIB_SIZE},
 };
-use freight_vm::TypeSystem;
+use freight_vm::{expression::Expression, TypeSystem};
 use std::{collections::HashMap, path::PathBuf, rc::Rc};
 
 #[derive(PartialEq, Eq, Clone, Debug)]
@@ -22,6 +23,12 @@ impl TypeSystem for FenderTypeSystem {
     type TypeId = FenderTypeId;
     type Init = FenderInitializer;
     type GlobalContext = FenderGlobalContext;
+}
+
+impl From<FenderValue> for Expression<FenderTypeSystem> {
+    fn from(value: FenderValue) -> Self {
+        Expression::RawValue(value.into())
+    }
 }
 
 #[derive(Default, Debug)]
